#15 RecordingActivity 패키지 분리
