--- app/src/main/java/com/example/androidlab2/RecordingItem.kt
@@ -1,3 +1,0 @@
-package com.example.androidlab2
-
-data class RecordingItem(val name: String, val date: Long)
