+++ app/src/main/java/com/example/androidlab2/model/RecordingItem.kt
@@ -1,0 +1,9 @@
+package com.example.androidlab2.model
+
+import java.io.File
+
+data class RecordingItem (
+    val name: String,
+    val date: Long,
+    val file: File
+)
